+++ ETA-SDK/API/Model/ETA_Branding.h
@@ -11,12 +11,10 @@
 @interface ETA_Branding : MTLModel<MTLJSONSerializing>
 
 @property (nonatomic, readwrite, strong) NSString* name;
-@property (nonatomic, readwrite, strong) NSString* urlName;
 @property (nonatomic, readwrite, strong) NSURL* websiteURL;
 @property (nonatomic, readwrite, strong) UIColor* color;
 @property (nonatomic, readwrite, strong) NSURL* logoURL;
-@property (nonatomic, readwrite, strong) UIColor* logoBackgroundColor;
 @property (nonatomic, readwrite, strong) NSURL* pageflipLogoURL;
 @property (nonatomic, readwrite, strong) UIColor* pageflipColor;
-
+@property (nonatomic, readwrite, strong) NSString* brandDescription;
 @end