+++ ETA-SDK/API/Model/ETA_Branding.m
@@ -20,12 +20,14 @@
 + (NSDictionary *)JSONKeyPathsByPropertyKey
 {
     return @{ @"name": @"name",
-              @"urlName": @"url_name",
               @"websiteURL": @"website",
+              @"brandDescription": @"description",
+              
               @"color": @"color",
               @"logoURL": @"logo",
+              
               @"pageflipLogoURL": @"pageflip.logo",
-              @"pageflipLogoColor": @"pageflip.color",
+              @"pageflipColor": @"pageflip.color",
               };
 }
 
@@ -37,9 +39,6 @@
 }
 + (NSValueTransformer *) logoURLJSONTransformer {
     return [NSValueTransformer valueTransformerForName:MTLURLValueTransformerName];
-}
-+ (NSValueTransformer *) logoBackgroundColorJSONTransformer {
-    return [NSValueTransformer valueTransformerForName:ETA_HexColor_ValueTransformerName];
 }
 + (NSValueTransformer *) pageflipLogoURLJSONTransformer {
     return [NSValueTransformer valueTransformerForName:MTLURLValueTransformerName];