+++ ETA-SDK/CatalogReader/ETA_CatalogReaderView.h
@@ -30,6 +30,8 @@ typedef NS_ENUM(NSInteger, ETA_CatalogReader_ErrorCode) {
 - (void) catalogReaderViewDidFinishFetchingData:(ETA_CatalogReaderView *)catalogReaderView error:(NSError*)error;
 
 
+- (void) catalogReaderView:(ETA_CatalogReaderView *)catalogReaderView didBeginTouchingLocation:(CGPoint)tapLocation onPageIndex:(NSUInteger)pageIndex hittingHotspots:(NSArray*)hotspots;
+- (void) catalogReaderView:(ETA_CatalogReaderView *)catalogReaderView didFinishTouchingLocation:(CGPoint)tapLocation onPageIndex:(NSUInteger)pageIndex hittingHotspots:(NSArray*)hotspots;
 - (void) catalogReaderView:(ETA_CatalogReaderView *)catalogReaderView didTapLocation:(CGPoint)tapLocation onPageIndex:(NSUInteger)pageIndex hittingHotspots:(NSArray*)hotspots;
 - (void) catalogReaderView:(ETA_CatalogReaderView *)catalogReaderView didLongPressLocation:(CGPoint)longPressLocation onPageIndex:(NSUInteger)pageIndex hittingHotspots:(NSArray*)hotspots;