+++ ETA-SDK/ETA.m
@@ -17,15 +17,6 @@ NSString* const ETA_AttachedUserChangedNotification = @"ETA_AttachedUserChangedN
 
 NSString* const ETA_AvoidServerCallKey = @"ETA_AvoidServerCallKey";
 
-NSString* const ETA_APIErrorDomain = @"ETA_APIErrorDomain";
-NSString* const ETA_APIError_URLResponseKey = @"ETA_APIError_URLResponseKey";
-NSString* const ETA_APIError_ErrorIDKey = @"ETA_APIError_IDKey";
-NSString* const ETA_APIError_ErrorObjectKey = @"ETA_APIError_ErrorObjectKey";
-
-typedef enum {
-    ETA_APIErrorCode_MissingParameter = 0
-} ETA_APIErrorCode;
-
 @interface ETA ()
 
 @property (nonatomic, readwrite, strong) ETA_APIClient* client;
@@ -177,9 +168,7 @@ static ETA* ETA_SingletonSDK = nil;
     if (!email || !password)
     {
         if (completionHandler)
-            completionHandler([NSError errorWithDomain:ETA_APIErrorDomain
-                                                  code:ETA_APIErrorCode_MissingParameter
-                                              userInfo:@{NSLocalizedDescriptionKey: @"Email and Password required"}]);
+            completionHandler([NSError SGN_errorWithSDKCode:SGN_SDKError_MissingParameter message:@"Email and Password required"]);
         return;
     }
     [self.client attachUser:@{@"email":email, @"password":password} withCompletion:completionHandler];
@@ -193,9 +182,7 @@ static ETA* ETA_SingletonSDK = nil;
     if (!facebookToken)
     {
         if (completionHandler)
-            completionHandler([NSError errorWithDomain:ETA_APIErrorDomain
-                                                  code:ETA_APIErrorCode_MissingParameter
-                                              userInfo:@{NSLocalizedDescriptionKey: @"Facebook token required"}]);
+            completionHandler([NSError SGN_errorWithSDKCode:SGN_SDKError_MissingParameter message:@"Facebook token required"]);
         return;
     }
     [self.client attachUser:@{@"facebook_token":facebookToken} withCompletion:completionHandler];
@@ -503,74 +490,6 @@ static ETA* ETA_SingletonSDK = nil;
     self.isLocationFromSensor = isFromSensor;
 }
 
-
-
-#pragma mark - Errors
-
-+ (NSDictionary*) errors
-{
-    static NSDictionary* errors = nil;
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
-        
-        errors = @{
-                   // Session errors
-                   @(1100): @"ETA_Error_SessionError",
-                   @(1101): @"ETA_Error_SessionTokenExpired",
-                   @(1102): @"ETA_Error_SessionInvalidAPIKey",
-                   @(1103): @"ETA_Error_SessionMissingSignature",
-                   @(1104): @"ETA_Error_SessionInvalidSignature",
-                   @(1105): @"ETA_Error_SessionTokenNotAllowed",
-                   @(1106): @"ETA_Error_SessionMissingOrigin",
-                   @(1107): @"ETA_Error_SessionMissingToken",
-                   @(1108): @"ETA_Error_SessionInvalidToken",
-                   
-                   
-                   // Authentication
-                   @(1200): @"ETA_Error_AuthenticationError",
-                   @(1201): @"ETA_Error_AuthenticationInvalidCredentials",
-                   @(1202): @"ETA_Error_AuthenticationNoUser",
-                   @(1203): @"ETA_Error_AuthenticationEmailNotVerified",
-                   
-                   
-                   // Authorization
-                   @(1300): @"ETA_Error_AuthorizationError",
-                   @(1301): @"ETA_Error_AuthorizationActionNotAllowed",
-                   
-                   
-                   // Missing information
-                   @(1400): @"ETA_Error_InfoRequestInvalid",
-                   @(1401): @"ETA_Error_InfoMissingGeolocation",
-                   @(1402): @"ETA_Error_InfoMissingRadius",
-                   @(1411): @"ETA_Error_InfoMissingAuthentication",
-                   @(1431): @"ETA_Error_InfoMissingEmail",
-                   @(1432): @"ETA_Error_InfoMissingBirthday",
-                   @(1433): @"ETA_Error_InfoMissingGender",
-                   @(1434): @"ETA_Error_InfoMissingLocale",
-                   @(1435): @"ETA_Error_InfoMissingName",
-                   @(1440): @"ETA_Error_InfoResourceNotFound",
-                   
-                   // Invalid information
-                   @(1500): @"ETA_Error_InfoInvalid",
-                   @(1501): @"ETA_Error_InfoInvalidResourceID",
-                   @(1530): @"ETA_Error_InfoResourceDuplication",
-                   @(1566): @"ETA_Error_InfoInvalidBodyData",
-                   
-                   // Internal corruption of data
-                   @(2000): @"ETA_Error_InternalIntegrityError",
-                   @(2010): @"ETA_Error_InternalSearchError",
-                   @(2015): @"ETA_Error_InternalNonCriticalError",
-                   
-                   // Misc.
-                   @(4000): @"ETA_Error_MiscActionNotExists",
-                   };
-    });
-    return errors;
-}
-+ (NSString*) errorForCode:(NSInteger)errorCode
-{
-    return [[ETA errors] objectForKey:@(errorCode)];
-}